fix occasional deadlock and data loss
This clarifies state transitions a little more, and adds more logging.
Also, this fixes the chief case of data loss in the protocol: packet
truncation at the point of recvmsg(), because GetUDPMTU() gave an answer
that was too low in some cases. IP layer packet fragmentation made up
for UDP packets over the MTU, but that won't save us if the packets then
get truncated because of a too-small receive buffer.

We will still need to address path MTU discovery in some better way,
ideally with rfc 4821 PLPMTUD.