make receive buffer larger than advertised
After looking carefully through the libutp code, I've verified that we
_should_ expect to receive more data than the receive buffer said it
could hold, even with a well-behaved peer, and that we can't easily drop
such excess packets to make them be transmitted again. Such excess data
should be infrequent (e.g., when the receive window has been 0 for >15
seconds, allow one extra packet) so doubling the actual buffer size
should make it less likely to have real overflows in normal conditions.

Of course, an overflow is still _possible_, especially with a malicious
peer, so for now we will return an error and shut down the connection.