replace logr with zap entirely (#9)
I thought logr would be a good way to avoid depending on a single
particular logging framework. But it has been too klunky to use,
due to its lowest-common-denominator approach, and it is inefficient,
showing up too prominently in profiling data.

Zap is optimized for speed, especially in the case where logging is
disabled, and in the worst case, it's possible to make a zap log handler
that passes info on to a different log framework, such as logrus.

I also took this opportunity to convert the printf-style debug
statements inherited from the C++ library into zap-style log statements.

I'm hoping the change in performance characteristics might be enough to
make it easier to reproduce the flow control problem we've been seeing.